+++ src/helpers/index.js
@@ -1,4 +1,4 @@
 import getElementPosition from './getElementPosition';
+import buildMessage from './buildMessage';
 
-
-export { getElementPosition };
+export { getElementPosition, buildMessage };
